bigger pass at changes, fixed broken specs
