+++ app/models/collaborator_solution.rb
@@ -1,0 +1,2 @@
+class CollaboratorSolution < Solution
+end
