+++ app/models/community_solution.rb
@@ -1,0 +1,3 @@
+class CommunitySolution < Solution
+  publishable
+end
