+++ app/representers/api/v1/collaborator_solution_representer.rb
@@ -1,0 +1,4 @@
+module Api::V1
+  class CollaboratorSolutionRepresenter < SolutionRepresenter
+  end
+end
