+++ app/representers/api/v1/community_solution_representer.rb
@@ -1,0 +1,7 @@
+module Api::V1
+  class CommunitySolutionRepresenter < SolutionRepresenter
+
+    publishable
+
+  end
+end
