+++ app/representers/api/v1/exercise_representer.rb
@@ -3,7 +3,9 @@ module Api::V1
 
     # Attachments may (for a while) contain collaborator solution attachments, so
     # only show them to those who can see solutions
-    has_attachments(if: lambda { |args| can_view_solutions?(args[:user]) })
+    has_attachments(
+      if: ->(args) { args.members.include?(:user) && can_view_solutions?(args[:user]) }
+    )
 
     has_logic
     has_tags
