+++ app/representers/api/v1/solution_representer.rb
@@ -1,10 +1,17 @@
 module Api::V1
-  class SolutionRepresenter < PublicationRepresenter
+  class SolutionRepresenter < Roar::Decorator
+
+    include Roar::JSON
 
     has_logic
     has_attachments
 
     property :title,
+             type: String,
+             writeable: true,
+             readable: true
+
+    property :type,
              type: String,
              writeable: true,
              readable: true
@@ -19,11 +26,6 @@ module Api::V1
              type: String,
              writeable: true,
              readable: true
-
-    property :is_by_collaborator,
-             type: 'boolean',
-             readable: true,
-             writeable: false
 
   end
 end
