--- app/representers/api/v1/solution_search_representer.rb
@@ -1,9 +1,0 @@
-module Api::V1
-  class SolutionSearchRepresenter < OpenStax::Api::V1::AbstractSearchRepresenter
-
-    collection :items, inherit: true,
-                       class: Solution,
-                       decorator: SolutionRepresenter
-
-  end
-end
