--- db/migrate/20160220021713_add_is_by_collaborator_to_solutions.rb
@@ -1,6 +1,0 @@
-class AddIsByCollaboratorToSolutions < ActiveRecord::Migration
-  def change
-    add_column :solutions, :is_by_collaborator, :boolean, null: false, default: false
-    # TODO migrate existing solutions to be by collaborator
-  end
-end
