+++ db/migrate/20160220021713_add_type_to_solutions.rb
@@ -1,0 +1,5 @@
+class AddTypeToSolutions < ActiveRecord::Migration
+  def change
+    add_column :solutions, :type, :string, null: false, default: "CollaboratorSolution"
+  end
+end
