+++ spec/factories/solutions.rb
@@ -1,5 +1,11 @@
 FactoryGirl.define do
-  factory :solution do
+  factory :collaborator_solution do
+    question
+    solution_type SolutionType::EXAMPLE
+    content "Worked example!"
+  end
+
+  factory :community_solution do
     question
     solution_type SolutionType::EXAMPLE
     content "Worked example!"
