+++ spec/models/question_spec.rb
@@ -4,7 +4,8 @@ RSpec.describe Question, type: :model do
 
   it { is_expected.to have_many(:stems).dependent(:destroy) }
   it { is_expected.to have_many(:answers).dependent(:destroy) }
-  it { is_expected.to have_many(:solutions).dependent(:destroy) }
+  it { is_expected.to have_many(:collaborator_solutions).dependent(:destroy) }
+  it { is_expected.to have_many(:community_solutions).dependent(:destroy) }
 
   it { is_expected.to have_many(:hints).dependent(:destroy) }
