--- spec/representers/api/v1/solution_search_representer.rb
@@ -1,7 +1,0 @@
-require 'rails_helper'
-
-module Api::V1
-  RSpec.describe SolutionSearchRepresenter do
-    pending "add some examples to (or delete) #{__FILE__}"
-  end
-end
