+++ spec/routing/api/v1/publications_routing_spec.rb
@@ -6,7 +6,7 @@ module Api::V1
 
       it "routes to #publish" do
         { 'exercises' => { :exercise_id => '1' },
-          'solutions' => { :solution_id => '1' }
+          'community_solutions' => { :community_solution_id => '1' }
         }.each do |path, params|
           expect(put("/api/#{path}/1/publish")).to(
             route_to("api/v1/publications#publish",
